+++ backend/src/main/java/team15/airbnb/accommodation/domain/AccommodationType.java
@@ -1,4 +1,4 @@
-package team15.airbnb.domain.accommodation;
+package team15.airbnb.accommodation.domain;
 
 public enum AccommodationType {
 	RESIDENCE, HOTEL, MOTEL, GUEST_HOUSE
