+++ backend/src/main/java/team15/airbnb/accommodation/domain/Address.java
@@ -1,4 +1,4 @@
-package team15.airbnb.domain.accommodation;
+package team15.airbnb.accommodation.domain;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
