+++ backend/src/main/java/team15/airbnb/accommodation/domain/Country.java
@@ -1,4 +1,4 @@
-package team15.airbnb.domain.accommodation;
+package team15.airbnb.accommodation.domain;
 
 public enum Country {
     SOUTH_KOREA, JAPAN, CHINA
