+++ backend/src/main/java/team15/airbnb/reservation/domain/Reservation.java
@@ -1,5 +1,7 @@
-package team15.airbnb.domain;
+package team15.airbnb.reservation.domain;
 
+import javax.validation.constraints.NotNull;
+import java.time.LocalDateTime;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -8,20 +10,32 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import team15.airbnb.domain.accommodation.Accommodation;
+import team15.airbnb.accommodation.domain.Accommodation;
+import team15.airbnb.common.domain.BaseEntity;
+import team15.airbnb.user.domain.User;
 
 @Entity
-public class Favorite {
+public class Reservation extends BaseEntity {
 
 	@Id
-	@Column(name = "favorite_id")
+	@Column(name = "reservation_id")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
+	@NotNull
+	@Column(columnDefinition = "TIMESTAMP")
+	private LocalDateTime checkInDate;
+
+	@NotNull
+	@Column(columnDefinition = "TIMESTAMP")
+	private LocalDateTime checkOutDate;
+
+	@NotNull
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "user_id")
 	private User user;
 
+	@NotNull
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "accommodation_id")
 	private Accommodation accommodation;
