+++ backend/src/main/java/team15/airbnb/user/domain/Favorite.java
@@ -1,7 +1,5 @@
-package team15.airbnb.domain;
+package team15.airbnb.user.domain;
 
-import javax.validation.constraints.NotNull;
-import java.time.LocalDateTime;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -10,30 +8,21 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import team15.airbnb.domain.accommodation.Accommodation;
+import team15.airbnb.accommodation.domain.Accommodation;
+import team15.airbnb.user.domain.User;
 
 @Entity
-public class Reservation extends BaseEntity{
+public class Favorite {
 
 	@Id
-	@Column(name = "reservation_id")
+	@Column(name = "favorite_id")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
-	@NotNull
-	@Column(columnDefinition = "TIMESTAMP")
-	private LocalDateTime checkInDate;
-
-	@NotNull
-	@Column(columnDefinition = "TIMESTAMP")
-	private LocalDateTime checkOutDate;
-
-	@NotNull
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "user_id")
 	private User user;
 
-	@NotNull
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "accommodation_id")
 	private Accommodation accommodation;
