+++ backend/src/main/java/team15/airbnb/user/domain/UserType.java
@@ -1,4 +1,4 @@
-package team15.airbnb.domain;
+package team15.airbnb.user.domain;
 
 public enum UserType {
 	CUSTOMER, HOST, SUPER_HOST
